pyproject.toml
[tool.poetry]
name = "uilibspec-binding-generator"
version = "0.1.0"
description = ""
authors = ["Maxime Bouroumeau-Fuseau <maxime.bouroumeau@gmail.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/hyperflask/uilibspec-binding-generator"
packages = [{include = "uilibspec_binding_generator", from="src"}]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0.1"
beautifulsoup4 = "^4.12.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"